+++ src/events/misc.h
@@ -6,5 +6,7 @@ void event_allaround_extract_remove_colon (struct irc_message_compo *);
 void event_local_and_global_users         (struct irc_message_compo *);
 void event_nicknameInUse                  (struct irc_message_compo *);
 void event_channel_forward                (struct irc_message_compo *);
+void event_channelModeIs                  (struct irc_message_compo *);
+void event_channelCreatedWhen             (struct irc_message_compo *);
 
 #endif